Remove Aasm::Project module.
